@@ -10,9 +10,9 @@ module.exports = {
     },
   ],
   format: "po",
-  compileNamespace: "cjs",
   sourceLocale: "de",
   fallbackLocales: {
     default: "de",
   },
+  compileNamespace: "ts",
 };
